Created veloverlet_1000.py
I started to write the python-file for the execution of 1000 steps of the velocity-verlet algorithm. It takes in the calibration data, generates a random starting sample as for now, and does 1000 time steps using the velocity-verlet algorithm. In every 10th of these steps, a list of all the nn-distances will be saved in or appended to an external file, which in turn can be used to calculate the pair-correlation-function using the python file like discussed